+++ assets/js/src/core/modules/element/dynamic-types/definitions/document/editable/hooks/use-pending-elements-reveal.tsx
